Just kidding on previous commit. Now rusanov flux kernel working.

+++ src/faces_kernels.cu
@@ -463,7 +463,7 @@ void apply_bcs_wrapper(mdvector_gpu<double> U, unsigned int nFpts, unsigned int 
 __global__
 void rusanov_flux(mdvector_gpu<double> U, mdvector_gpu<double> Fconv, 
     mdvector_gpu<double> Fcomm, mdvector_gpu<double> P, mdvector_gpu<double> norm,
-    mdvector_gpu<double>outnorm, mdvector_gpu<double> waveSp, double gamma, double rus_k,
+    mdvector_gpu<int> outnorm, mdvector_gpu<double> waveSp, double gamma, double rus_k,
     unsigned int nFpts, unsigned int nVars, unsigned int nDims)
 {
   const unsigned int fpt = blockDim.x * blockIdx.x + threadIdx.x;
@@ -539,7 +539,7 @@ void rusanov_flux(mdvector_gpu<double> U, mdvector_gpu<double> Fconv,
 
 void rusanov_flux_wrapper(mdvector_gpu<double> U, mdvector_gpu<double> Fconv, 
     mdvector_gpu<double> Fcomm, mdvector_gpu<double> P, mdvector_gpu<double> norm,
-    mdvector_gpu<double>outnorm, mdvector_gpu<double> waveSp, double gamma, double rus_k,
+    mdvector_gpu<int> outnorm, mdvector_gpu<double> waveSp, double gamma, double rus_k,
     unsigned int nFpts, unsigned int nVars, unsigned int nDims)
 {
   unsigned int threads = 192;
